Replace lorempixel.com with unsplash.it
The live carousel documentation was emitting mixed content warnings
because the protocol was hard-coded to http. Changing the image paths
to be protocol-agnostic did not work because lorempixel.com's cert
has been expired for more than a year.

It seems to me that putting chosen images in the asset path would be
a better long-term fix, but this should eliminate the warnings for
now.